refactor - moving things into modules
