+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__/
 .DS_Store
 *.py.swp
+.env
