+++ opds_abs/__init__.py
@@ -1,0 +1,5 @@
+"""
+OPDS Feed for Audiobookshelf
+"""
+
+__version__ = '0.1.0'
