+++ opds_abs/api/__init__.py
@@ -1,0 +1,5 @@
+"""API client for Audiobookshelf"""
+
+from opds_abs.api.client import fetch_from_api, get_download_urls_from_item
+
+__all__ = ['fetch_from_api', 'get_download_urls_from_item']
