+++ opds_abs/core/__init__.py
@@ -1,0 +1,6 @@
+"""Core functionality for the OPDS feed"""
+
+from opds_abs.core.feed_generator import BaseFeedGenerator
+from opds_abs.core.navigation import navigation
+
+__all__ = ['BaseFeedGenerator', 'navigation']
