+++ opds_abs/utils/__init__.py
@@ -1,0 +1,4 @@
+"""Utility functions for the OPDS feed"""
+
+# This will be populated as utility functions are added
+__all__ = []
