+++ run.py
@@ -1,0 +1,8 @@
+"""
+OPDS Feed for Audiobookshelf - Entry Point
+"""
+import uvicorn
+from opds_abs.main import app
+
+if __name__ == "__main__":
+    uvicorn.run("opds_abs.main:app", host="0.0.0.0", port=8000, reload=True)
